@@ -9,24 +9,25 @@
     "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
     "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
     "test:unit": "vitest",
-    "lint": "prettier --plugin-search-dir . --check .",
-    "format": "prettier --plugin-search-dir . --write ."
+    "lint": "prettier --check .",
+    "format": "prettier --write ."
   },
   "devDependencies": {
-    "@sveltejs/adapter-auto": "^2.0.0",
-    "@sveltejs/adapter-node": "^1.2.4",
-    "@sveltejs/kit": "^1.5.0",
+    "@sveltejs/adapter-auto": "^6.0.1",
+    "@sveltejs/adapter-node": "^5.2.12",
+    "@sveltejs/kit": "^2.21.1",
+    "@sveltejs/vite-plugin-svelte": "^5.0.0",
     "dayjs": "^1.11.7",
     "grammy": "^1.16.0",
     "pino": "^8.17.2",
-    "prettier": "^2.8.0",
-    "prettier-plugin-svelte": "^2.8.1",
-    "svelte": "^3.54.0",
-    "svelte-check": "^3.0.1",
+    "prettier": "^3.0.0",
+    "prettier-plugin-svelte": "^3.4.0",
+    "svelte": "^5.33.0",
+    "svelte-check": "^4.2.1",
     "tslib": "^2.4.1",
     "typescript": "^5.0.0",
-    "vite": "^4.3.0",
-    "vitest": "^0.25.3"
+    "vite": "^6.0.0",
+    "vitest": "^3.1.4"
   },
   "type": "module"
 }
